+++ src/Futhark/CodeGen/Backends/CCUDA.hs
@@ -52,7 +52,8 @@ cliOptions :: [Option]
 cliOptions = [ Option { optionLongName = "dump-cuda"
                       , optionShortName = Nothing
                       , optionArgument = RequiredArgument "FILE"
-                      , optionAction = [C.cstm|futhark_context_config_dump_program_to(cfg, optarg);|]
+                      , optionAction = [C.cstm|{futhark_context_config_dump_program_to(cfg, optarg);
+                                                entry_point = NULL;}|]
                       }
              , Option { optionLongName = "load-cuda"
                       , optionShortName = Nothing
@@ -62,12 +63,18 @@ cliOptions = [ Option { optionLongName = "dump-cuda"
              , Option { optionLongName = "dump-ptx"
                       , optionShortName = Nothing
                       , optionArgument = RequiredArgument "FILE"
-                      , optionAction = [C.cstm|futhark_context_config_dump_ptx_to(cfg, optarg);|]
+                      , optionAction = [C.cstm|{futhark_context_config_dump_ptx_to(cfg, optarg);
+                                                entry_point = NULL;}|]
                       }
              , Option { optionLongName = "load-ptx"
                       , optionShortName = Nothing
                       , optionArgument = RequiredArgument "FILE"
                       , optionAction = [C.cstm|futhark_context_config_load_ptx_from(cfg, optarg);|]
+                      }
+             , Option { optionLongName = "nvrtc-option"
+                      , optionShortName = Nothing
+                      , optionArgument = RequiredArgument "OPT"
+                      , optionAction = [C.cstm|futhark_context_config_add_nvrtc_option(cfg, optarg);|]
                       }
              , Option { optionLongName = "print-sizes"
                       , optionShortName = Nothing